% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_project}
\alias{check_project}
\title{Check integrity of project}
\usage{
check_project()
}
\description{
The purpose of \code{check_project()} is to be for production projects what
\verb{R CMD check} is for packages. A testing suite that verifies that everything
is as it should be to the best of its abilities.
}
