% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_production}
\alias{create_production}
\title{Create a new production project}
\usage{
create_production(
  path,
  type = c("batch", "app", "api", "report"),
  ...,
  open = rlang::is_interactive(),
  report_type = "default"
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{type}{A string giving the type of project to create. Currently \code{"batch"},
\code{"app"}, \code{"api"}, and \code{"report"} is recognized.}

\item{...}{
  Arguments passed on to \code{\link[usethis:create_package]{usethis::create_project}}
  \describe{
    \item{\code{rstudio}}{If \code{TRUE}, calls \code{\link[usethis:use_rstudio]{use_rstudio()}} to make the new package or
project into an \href{https://r-pkgs.org/workflow101.html#sec-workflow101-rstudio-projects}{RStudio Project}.
If \code{FALSE} and a non-package project, a sentinel \code{.here} file is placed so
that the directory can be recognized as a project by the
\href{https://here.r-lib.org}{here} or
\href{https://rprojroot.r-lib.org}{rprojroot} packages.}
  }}

\item{open}{If \code{TRUE}, \link[usethis:proj_activate]{activates} the new project:
\itemize{
\item If using RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}

\item{report_type}{If \code{type} is \code{"report"}, what kind of report}
}
\value{
Path to the newly created project, invisibly.
}
\description{
This function works like \code{\link[usethis:create_package]{usethis::create_project()}}, but extends the
behavior to match the setup advocated by producethis, most importantly the
existance of a \code{DESCRIPTION} file with the \code{Type} field set to the type of
production job this project will be.
}
