% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{deploy_repo_to_connect}
\alias{deploy_repo_to_connect}
\title{Deploys and updates project on Connect}
\usage{
deploy_repo_to_connect(branch = "gh-connect")
}
\arguments{
\item{branch}{The branch to deploy from. It is expected that it exists and
contain the necessary files for Connect}
}
\value{
This function is called for its sideeffects
}
\description{
This function is called once the project is ready for deployment, having
already called \code{\link[=prepare_for_connect]{prepare_for_connect()}} and pushed to the deployment branch.
It is highly recommended that you do not deploy from your local machine but
rather from a versioned repository such as GitHub.
}
