% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{prepare_for_gha}
\alias{prepare_for_gha}
\title{Prepares a production project for deployment on GitHub Actions}
\usage{
prepare_for_gha()
}
\value{
This function is called for its sideeffects
}
\description{
This function generates all the necessary files for your project to be
executed on GitHub Actions. It is expected that this function
is not called interactively, but rather as part of an automation using e.g.
GitHub Actions. This function will create an action
}
