% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producethis-package.R
\docType{package}
\name{producethis-package}
\alias{producethis}
\alias{producethis-package}
\title{producethis: Set Up Commonly Used 🏭 Components}
\description{
Automate tasks related to putting R code in production. Based on an opinionated folder structure it provides a variety of helpers for testing, deploying, and running R code in production.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/producethis}
  \item Report bugs at \url{https://github.com/r-lib/producethis/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
