% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_access.R
\name{use_access}
\alias{use_access}
\title{Set access rights on your project}
\usage{
use_access(users = NULL, groups = NULL, global = NULL)
}
\arguments{
\item{users}{A character vector of user names that should have access to the
deployment}

\item{groups}{A character vector of group names that should have access to
deployment}

\item{global}{A global setting for permissions. For Posit Connect this can
be \code{"all"} or \code{"logged-in"}.}
}
\value{
The function is called for its side effects.
}
\description{
This function is used to record the permission levels for the deployment.
The validity of the given values depend on the deployment type and the
users/groups that have been registered there. It will only be checked at
deployment time. In general, permission is understood as "viewing right",
since editing all should happen through the git repository anyway.
}
