% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_envvars.R
\name{use_envvars}
\alias{use_envvars}
\alias{use_local_envvars}
\alias{set_envvars}
\title{Add one or more environment variables to the project}
\usage{
use_envvars(...)

use_local_envvars(...)

set_envvars()
}
\arguments{
\item{...}{name-value pairs giving the name and value of the environment
variable(s) to set}
}
\value{
These functions are called for their side effects
}
\description{
Environment variables can be used to control various aspects of the execution
of a production job. For environment variables that does not contain any
secret information it is beneficial to have them stored in the project to
make sure that they are set equally in all environments they are executed.
\code{use_envvars} will write the given environment variables to the \code{Envvar} field
of the DESCRIPTION file. During deployment these values are forwarded to the
deployment server. \code{set_envvars()} will register the environment variables in
the current R session for interactive use (this happens automatically in
\code{\link[=try_connect_run]{try_connect_run()}}).
}
\note{
\strong{NEVER} store secret/confidential information in this way. This is
purely for use for behavior changing environment variables.
}
