% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{use_github_gha}
\alias{use_github_gha}
\title{Deploy a project using GitHub actions}
\usage{
use_github_gha(
  from = c("main", "master"),
  result_branch = "gh-results",
  output_dir = "output"
)
}
\arguments{
\item{from}{A character vector giving which branch(es) to publish from}

\item{result_branch}{The branch to push the content of \code{output_dir} to after
the execution ends}

\item{output_dir}{The directory any output from the execution is expected to
end up in. Do note that it is the responsibility of the production code to
actually write to this folder}
}
\value{
This function is called for its side effects
}
\description{
While all deployments based on producethis uses GitHub actions, this
deployment type uses it exclusively, meaning that the code is also executed
in an action. This puts some limits on the kind of project that can be
deployed, since GHA doesn't provide a web server. Thus, only batch and static
reports can be deployed in this way.
}
