% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_schedule.R
\name{use_schedule}
\alias{use_schedule}
\alias{at}
\alias{interval}
\title{Execute a deployed project on a schedule}
\usage{
use_schedule(
  minute = NULL,
  hour = NULL,
  day = NULL,
  weekday = NULL,
  week = NULL,
  month = NULL,
  year = NULL
)

at(x = NULL, to = NULL, step = NULL)

interval(x)
}
\arguments{
\item{minute, hour, day, weekday, week, month, year}{Scheduling times. See Datails
on how to set them}

\item{x}{An interval or time location}

\item{to}{The end of the schedule range}

\item{step}{The step size of the schedule range}
}
\description{
For projects that does not run continuously, such as APIs or apps, it may
make sense to execute them periodically rather than only once when deployed
or manually when you trigger it. This function records an intended execution
schedule in the DESCRIPTION file that will be set during deployment. Do note
that not all deployment environments support all scheduling settings. If it
is deducible what type of deployment you are using the provided schedule will
be tested before recording it.
}
