% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_settings.R
\name{use_settings}
\alias{use_settings}
\alias{use_connect_settings}
\alias{use_r_settings}
\alias{use_gha_settings}
\title{Provide additional settings for deployment}
\usage{
use_connect_settings(...)

use_r_settings(...)

use_gha_settings(...)
}
\arguments{
\item{...}{key-value pairs of settings. See details for allowed names. Set
a key to \code{NULL} to remove it and thus use the default.}
}
\value{
This function is called for it's sideeffect
}
\description{
These functions will record additional deployment specific settings for the
project, on top of the settings that are otherwise recorded in other
DESCRIPTION fields.
}
\details{
Settings are specific to a given deployment target. Currently only Posit
Connect is supported.
}
\section{use_connect_settings()}{
Possible settings for Posit Connect are:
\itemize{
\item \code{locked}
\item \code{locked_message}
\item \code{connection_timeout}
\item \code{read_timeout}
\item \code{init_timeout}
\item \code{idle_timeout}
\item \code{max_processes}
\item \code{min_processes}
\item \code{max_conns_per_process}
\item \code{load_factor}
\item \code{cpu_request}
\item \code{cpu_limit}
\item \code{memory_request}
\item \code{memory_limit}
\item \code{amd_gpu_limit}
\item \code{nvidia_gpu_limit}
\item \code{default_image_name}
\item \code{default_r_environment_management}
\item \code{default_py_environment_management}
}

Please consult the API documentation for your Connect instance for
description and allowed values for the various settings.
}

